Fix(macos): Make scrolling work for the field view
When mouse is connected and scroll bars are set to always show
(System preferences -> General -> Show scroll bars) each field value
gets a vertical scrollbar regardless of being only 1 line long.

This scrollbar will move the value up and down by a couple of pixels.
It not only looks ugly, but also prevents overall table scrolling,
when the amount of fields exceeds the screen height. This is so,
because inner scrollbars receive focus and do not let the outer
scroll view to do the job