Refine: more detailed crash logs for memory protection issues
There are occasional crashes caused by the failure to decrypt SecureBytes content.
They should not happen at all, but they do. To learn more about the cause,
we need to know at least the error code returned by `SecKeyCreateDecryptedData`.

There seems to be no standard way to include error info in a crash log,
so as a workaround we include it by calling either `_bit_0` or `_bit_1` function
for each bit of the error code. This way, the error code will be reflected in the
crash stack trace.