Refine(premium): show unlimited DB timeout as free
Historically, DB timeout option `never` was initially intended as a premium feature.
However, this conflicted with the widely requested biometric unlock. So the "never"
option went into an odd trade-off state where it was _displayed_ as premium,
but _behaved_ as free. This caused quite a bit of user confusion, too.

Now it will both behave and appear as free.