Fix(macos): erase app settings after reinstallation
On iOS, the system erases app settings on uninstall. MacOS, however, does not do that,
so KeePassium survived reinstallation with all the data intact, like nothing happened.
Which could be a security issue.

With this commit, KeePassium properly detects its own reinstallation on macOS,
and cleans up the settings and keychain data accordingly.