Fix(autoFill): fallback to UI mode if need clipboard access (#242)
Quick AutoFill does not have access to clipboard, so it cannot copy TOTP of a matching entry.

This commit switches Quick AutoFill to "slow" UI-based mode, so it can copy the OTP to clipboard
and quickly finish up. Not perfect, but still better than manually typing the code from a notification.

[thanks, David]